refactoring project structure: zeroconf classes
