--- classes/PowerState/index.js
@@ -1,5 +1,0 @@
-const ChangeStateZeroconf = require('./ChangeStateZeroconf');
-
-module.exports = {
-  ChangeStateZeroconf,
-};
